In the overworld now!

+++ include/global.fieldmap.h
@@ -131,7 +131,7 @@ struct MapEvents
 struct MapConnection
 {
     u8 direction;
-    u32 offset;
+    u64 offset;
     u8 mapGroup;
     u8 mapNum;
 };
